Add "up" arrows to page headers
Required some associated changes to the design of other elements. Looks
pretty cool now. The page-header is made smaller on mobile so as to fit
more with the "mobiley" look of the rest of the site.

This commit also includes a fix for images in the contribute section
which would have been to wide otherwise (setting max-width).